requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
dj-database-url==2.1.0
Django==4.2.7
django-fernet-fields-v2==0.9
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
Brotli==1.1.0
python-decouple==3.8